rename to cnot error density
